W172k: two MATCHes + six window clears.
btlUnit 00288DA0 -> MATCH: three branch-fed assignments expressed as CONDITIONAL OPERATORS, which fixed the branch targets. It had been sitting at 480/480 nd3 for two passes; the ternary form was the missing piece.
nmCmdList 003C69D0 -> MATCH: reloaded counter fields, direct flag updates, preincremented counters.
nmCmdList 003C5FA0 248/288 nd163 -> 284/288 nd159 -- I verified the agent's unmeasured final edit (a u32 -> u8 local width at line 333) and it improves BOTH size and nd, so it stays.
fclShop 003E5E20 nd100 -> 96 via the equivalent upper-bound form '> 0x168' rather than '>= 0x169'.

itfMesManager FOUR clears -- 003A30C0, 003A37C0, 003A5980 (888/800 nd638 -> 796/800 nd364), plus 003A5570/003A56F0 from the earlier pass -- by recovering 32-bit pointer/handle/intermediate widths, HIDDEN free/count arguments, direct typed local calls, correct branch body ordering, and removing an extra callback dereference. comuSasoi 003C2C50 cleared by removing a needless persistent s16 temporary so the casts rematerialize at call sites without a saved register.

Slice 434 -> 437 MATCH, over-window 12 -> 6, zero lost, zero additions, itfMesManager LF-only preserved (0 CRLF, 6201 LF), annotated barriers and pragma brackets untouched.

+++ src/Yajima/y_runtime.c
@@ -13108,7 +13108,7 @@ void FUN_0044af60(u32 param_1)
     }
     FUN_001b0250_typed(*(u32 *)((u8 *)DAT_0086ef10 + *(char *)(iVar1 + 1) * 0x1c0),
                  *(u32 *)(iVar12 + 0x20));
-    lVar10 = FUN_001afa20_typed((*(u32 *)(iVar12 + 8)),(void*)(*(u32 *)((u8 *)DAT_0086ef10 + *(char *)(iVar1 + 1) * 0x1c0)),(void*)(iVar1 + 0x6c));
+    lVar10 = FUN_001afa20_typed((*(f32 *)(iVar12 + 8)),(void*)(*(u32 *)((u8 *)DAT_0086ef10 + *(char *)(iVar1 + 1) * 0x1c0)),(void*)(iVar1 + 0x6c));
     if (lVar10 == 0) {
       return;
     }
@@ -13162,7 +13162,7 @@ void FUN_0044af60(u32 param_1)
     break;
   case 7:
     FUN_001b0250_typed(*(u32 *)((u8 *)DAT_0086ef10 + iVar11),0);
-    lVar10 = FUN_001afa20_typed((*(u32 *)(iVar12 + 0xc)),(void*)(*(u32 *)((u8 *)DAT_0086ef10 + *(char *)(iVar1 + 1) * 0x1c0)),(void*)(iVar1 + 0x6c));
+    lVar10 = FUN_001afa20_typed((*(f32 *)(iVar12 + 0xc)),(void*)(*(u32 *)((u8 *)DAT_0086ef10 + *(char *)(iVar1 + 1) * 0x1c0)),(void*)(iVar1 + 0x6c));
     if (lVar10 != 0) {
       *(u32 *)(iVar1 + 0x60) = *(u32 *)(iVar1 + 0x6c);
       *(u32 *)(iVar1 + 100) = *(u32 *)(iVar1 + 0x70);
